requirements.txt
tornado
python-socketio
googletrans
